+++ utilities/extended_data_view_test.ts
@@ -1,8 +1,8 @@
-import { assertEquals } from "./deps.ts";
-import { BetterDataView } from "./utils.ts";
+import { assertEquals } from "../deps.ts";
+import { ExtendedDataView } from "./extended_data_view.ts";
 
-Deno.test("BetterDataView", async (t) => {
-  const dataView = new BetterDataView(
+Deno.test("ExtendedDataView", async (t) => {
+  const dataView = new ExtendedDataView(
     // deno-fmt-ignore
     new Uint8Array([0x5F, 0xB6, 0x4A, 0xEB, 0x4D, 0x65, 0xAE, 0xC9, 0x02, 0x05, 0xBD, 0x76, 0x55, 0x9A, 0xAF, 0x29]).buffer,
   );
@@ -20,7 +20,7 @@ Deno.test("BetterDataView", async (t) => {
   });
 
   await t.step("setBigUint128", () => {
-    const dataView = new BetterDataView(new Uint8Array(16).buffer);
+    const dataView = new ExtendedDataView(new Uint8Array(16).buffer);
     dataView.setBigUint128(0, 127223177168702480293718884489214209833n);
     assertEquals(
       dataView.getBigUint128(0, false),
